feat(lazy_hgraph): support extra info filtering (#2263)
* feat(lazy_hgraph): support extra info filtering


* feat(lazy_hgraph): support vector updates


* fix: align hgraph serialization types


* fix: address review feedback

Rename the LazyHGraph extra_info example to avoid reusing example number 323 and validate BruteForce UpdateVector inputs.


* fix: address LazyHGraph review and CI issues


---------

+++ src/algorithm/hgraph/hgraph.cpp
@@ -491,8 +491,8 @@ HGraph::GetStats() const {
     // index was not built from an imported cache.
     if (this->build_cache_hit_rate_ >= 0.0F) {
         stats["build_cache_hit_rate"].SetFloat(this->build_cache_hit_rate_);
-        stats["build_cache_hit_nodes"].SetInt(this->build_cache_hit_nodes_);
-        stats["build_cache_missed_nodes"].SetInt(this->build_cache_missed_nodes_);
+        stats["build_cache_hit_nodes"].SetUint64(this->build_cache_hit_nodes_);
+        stats["build_cache_missed_nodes"].SetUint64(this->build_cache_missed_nodes_);
     } else {
         stats["build_cache_hit_rate"]["skipped_reason"].SetString(
             "index was not built from an imported cache");
