+++ src/datacell/sparse_graph_datacell_parameter.cpp
@@ -25,7 +25,7 @@ SparseGraphDatacellParameter::SparseGraphDatacellParameter()
 void
 SparseGraphDatacellParameter::FromJson(const JsonType& json) {
     if (json.Contains(GRAPH_PARAM_MAX_DEGREE_KEY)) {
-        this->max_degree_ = json[GRAPH_PARAM_MAX_DEGREE_KEY].GetInt();
+        this->max_degree_ = json[GRAPH_PARAM_MAX_DEGREE_KEY].GetUint64();
     }
     if (json.Contains(GRAPH_SUPPORT_REMOVE)) {
         this->support_delete_ = json[GRAPH_SUPPORT_REMOVE].GetBool();
@@ -44,7 +44,7 @@ SparseGraphDatacellParameter::FromJson(const JsonType& json) {
 JsonType
 SparseGraphDatacellParameter::ToJson() const {
     JsonType json;
-    json[GRAPH_PARAM_MAX_DEGREE_KEY].SetInt(this->max_degree_);
+    json[GRAPH_PARAM_MAX_DEGREE_KEY].SetUint64(this->max_degree_);
     json[GRAPH_SUPPORT_REMOVE].SetBool(this->support_delete_);
     json[REMOVE_FLAG_BIT].SetInt(this->remove_flag_bit_);
     json[SUPPORT_DUPLICATE].SetBool(this->support_duplicate_);
